+++ src/app/modules/pages/editor/components/sidebar/layout-selector/layout-selector.component.ts
@@ -4,7 +4,8 @@ import {EMPTY, Observable} from "rxjs";
 import {Preset} from "@app/core/model/preset";
 import {toAbsoluteCMSUrl} from "@app/core/editor/utils";
 import {ActivatedRoute} from "@angular/router";
-import {mergeMap} from "rxjs/operators";
+import {distinctUntilChanged, map, mergeMap, take, tap} from "rxjs/operators";
+import {faVideo} from "@fortawesome/free-solid-svg-icons/faVideo";
 
 @Component({
   selector: 'app-layout-selector',
@@ -16,6 +17,9 @@ export class LayoutSelectorComponent implements OnInit {
   layout = new EventEmitter<Preset>();
 
   settings$: Observable<Preset[]> = EMPTY;
+  previewUrls: Map<number, string> = new Map<number, string>();
+
+  faVideo = faVideo;
 
   constructor(private cmsService: CmsService,
               private route: ActivatedRoute) {
@@ -23,7 +27,15 @@ export class LayoutSelectorComponent implements OnInit {
 
   ngOnInit(): void {
     this.settings$ = this.route.queryParamMap.pipe(
-      mergeMap(data => this.cmsService.getLayoutSetting(data.get('presetTag')))
+      map(param => param.get('presetTag')),
+      distinctUntilChanged(),
+      mergeMap(param => this.cmsService.getLayoutSetting(param)),
+      tap(presets => {
+        // Setting the preview urls and reading them in template
+        for (const preset of presets) {
+          this.previewUrls.set(preset.id, LayoutSelectorComponent.mapUrl(preset));
+        }
+      })
     );
   }
 
@@ -33,5 +45,22 @@ export class LayoutSelectorComponent implements OnInit {
 
   cmsApiUrl(url: string) {
     return toAbsoluteCMSUrl(url)
+  }
+
+  private static mapUrl(preset: Preset) {
+    const forcePreview = !preset.backgroundImage.mime.startsWith('image');
+    let url = forcePreview ? null : preset.backgroundImage?.formats?.thumbnail?.url;
+
+    if (preset.preview) {
+      url = preset.preview.formats?.thumbnail?.url;
+    }
+
+    if (!url) {
+      // if it is a video without preview url, then there should be an error loged
+      console.error('video preset', preset.id, 'must contain preview image');
+      return '';
+    }
+
+    return toAbsoluteCMSUrl(url);
   }
 }